Functional song selection and required choice
